(NOBIDS) Fix wrong data type on validator_statas columns

+++ db/migrations/20230411195211_add_proposer_statistic_columns.sql
@@ -1,9 +1,9 @@
 -- +goose Up
 -- +goose StatementBegin
 SELECT 'up SQL query';
-ALTER TABLE validator_stats ADD COLUMN IF NOT EXISTS cl_proposer_rewards_gwei INTEGER;
-ALTER TABLE validator_stats ADD COLUMN IF NOT EXISTS cl_proposer_rewards_gwei_total INTEGER;
-ALTER TABLE validator_performance ADD COLUMN IF NOT EXISTS cl_proposer_performance_total INTEGER;
+ALTER TABLE validator_stats ADD COLUMN IF NOT EXISTS cl_proposer_rewards_gwei BIGINT;
+ALTER TABLE validator_stats ADD COLUMN IF NOT EXISTS cl_proposer_rewards_gwei_total BIGINT;
+ALTER TABLE validator_performance ADD COLUMN IF NOT EXISTS cl_proposer_performance_total BIGINT;
 -- +goose StatementEnd
 
 -- +goose Down
